pyproject.toml
[project]
name = "llm-ideology-explorer"
version = "0.1.0"
description = "Educational toolkit demonstrating two-stage prompting methodology for analyzing ideological biases in LLMs"
authors = [{ name = "Sebastian Gomez", email = "sgomezgomez@gmail.com" }]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "jupyter>=1.0.0",
    "huggingface-hub>=0.20.0",
    "openai>=1.12.0",
    "rich>=10.0.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
pythonpath = [
  "src"
]

[tool.ruff]
line-length = 88